chore: update docs and refactor shared Criteria
